+++ picky/src/ssh/mod.rs
@@ -4,22 +4,26 @@ pub mod encode;
 pub mod private_key;
 pub mod public_key;
 
+use crate::key::{ec::NamedEcCurve, ed::NamedEdAlgorithm, EcCurve, EdAlgorithm, KeyError};
+
+use byteorder::ReadBytesExt;
+use std::io::{self, Read};
+
 pub use certificate::{SshCertKeyType, SshCertType, SshCertificate, SshCertificateBuilder};
 pub use private_key::SshPrivateKey;
 pub use public_key::SshPublicKey;
 
-use crate::key::{ec::NamedEcCurve, EcCurve, KeyError};
-use byteorder::ReadBytesExt;
-use std::io::{self, Read};
-
 pub(crate) type Base64Writer<'a, T, E> = base64::write::EncoderWriter<'a, T, E>;
 pub(crate) type Base64Reader<'a, T, E> = base64::read::DecoderReader<'a, T, E>;
 
+const SSH_COMBO_ED25519_KEY_LENGTH: usize = ed25519_dalek::SECRET_KEY_LENGTH + ed25519_dalek::PUBLIC_KEY_LENGTH;
+
 mod key_type {
-    pub const SSH_RSA: &str = "ssh-rsa";
+    pub const RSA: &str = "ssh-rsa";
     pub const ECDSA_SHA2_NIST_P256: &str = "ecdsa-sha2-nistp256";
     pub const ECDSA_SHA2_NIST_P384: &str = "ecdsa-sha2-nistp384";
     pub const ECDSA_SHA2_NIST_P521: &str = "ecdsa-sha2-nistp521";
+    pub const ED25519: &str = "ssh-ed25519";
 }
 
 mod key_identifier {
@@ -52,6 +56,24 @@ impl EcCurveSshExt for NamedEcCurve {
             // See comment inside function above
             NamedEcCurve::Unsupported(_) if self.is_nist_p521() => Ok(key_identifier::ECDSA_SHA2_NIST_P521),
             NamedEcCurve::Unsupported(oid) => Err(KeyError::unsupported_curve(oid, "ssh key identifier serialization")),
+        }
+    }
+}
+
+trait EdAlorithmSshExt {
+    fn to_ed_ssh_key_type(&self) -> Result<&'static str, KeyError>;
+}
+
+impl EdAlorithmSshExt for NamedEdAlgorithm {
+    fn to_ed_ssh_key_type(&self) -> Result<&'static str, KeyError> {
+        match self {
+            NamedEdAlgorithm::Known(EdAlgorithm::Ed25519) => Ok(key_type::ED25519),
+            NamedEdAlgorithm::Known(EdAlgorithm::X25519) => Err(KeyError::UnsupportedAlgorithm {
+                algorithm: "X25519 can't be use for SSH EdDSA keys",
+            }),
+            NamedEdAlgorithm::Unsupported(oid) => {
+                Err(KeyError::unsupported_ed_algorithm(oid, "ssh key type serialization"))
+            }
         }
     }
 }